backend/requirements.txt
Django>=3.0,<4.0
django-allauth==0.44
djangorestframework>=3.1,<4.0
djangorestframework-simplejwt>=4.6,<5.0
django-cors-headers>=3.7,<4.0
psycopg2>=2.9,<3.0
